requirements.txt
torch
transformers
datasets
sentencepiece
faiss-cpu
numpy
pandas
scikit-learn
nltk
python-docx
PyPDF2